% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{c.FeatureImportanceLearner}
\alias{c.FeatureImportanceLearner}
\title{Combine two FeatureImportanceLearner objects}
\usage{
\method{c}{FeatureImportanceLearner}(x, y, ...)
}
\arguments{
\item{x, y}{\verb{([FeatureImportanceLearner])} Objects to combine. Must have computed scores.}

\item{...}{(any) Ignored.}
}
\value{
New object of the same subclass as \code{x} and \code{y}.
}
\description{
Combine two FeatureImportanceLearner objects
}
\examples{
if (FALSE) {

}
}
