% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{c.FeatureImportanceMethod}
\alias{c.FeatureImportanceMethod}
\title{Combine two FeatureImportanceMethod objects}
\usage{
\method{c}{FeatureImportanceMethod}(x, y, ...)
}
\arguments{
\item{x, y}{\verb{([FeatureImportanceMethod])} Objects to combine. Must have computed scores.}

\item{...}{(any) Ignored.}
}
\value{
New object of the same subclass as \code{x} and \code{y}.
}
\description{
Combine two FeatureImportanceMethod objects
}
\examples{
\dontshow{if (requireNamespace("ranger", quietly = TRUE) && requireNamespace("mlr3learners", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(mlr3)
task = tgen("2dnormals")$generate(n = 100)

pfi1 = PFI$new(
  task = task,
  learner = lrn("classif.ranger", num.trees = 50, predict_type = "prob"),
  measure = msr("classif.ce"),
  features = "x1"
)
pfi1$compute()

pfi2 = PFI$new(
  task = task,
  learner = lrn("classif.ranger", num.trees = 50, predict_type = "prob"),
  measure = msr("classif.ce"),
  features = "x2"
)
pfi2$compute()

combined = c(pfi1, pfi2)
\dontshow{\}) # examplesIf}
}
