% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_score.R
\name{compute_score}
\alias{compute_score}
\title{Scoring utility for computing importances}
\usage{
compute_score(
  scores_pre,
  scores_post,
  relation = c("difference", "ratio"),
  minimize = TRUE
)
}
\arguments{
\item{scores_pre, scores_post}{(\code{numeric()}) Vector of scores or loss values at baseline / before (\verb{_pre}) a modification, and after (\verb{_post}) a modification (e.g., permutation or refit).}

\item{relation}{(\code{character(1)}: \code{"difference"}) Calculate the difference or \code{"ratio"} between pre and post modification value.}

\item{minimize}{(\code{logical(1)}: \code{TRUE}) Flips the order of the comparison, typically taken from a \link[mlr3:Measure]{mlr3::Measure}'s \verb{$minimize} field.}
}
\description{
Computes the \code{relation} of score before a change (e.g. PFI, LOCO, ...) and after.
If \code{minimize == TRUE}, then \code{scores_post - scores_pre} is computed for
\code{relation == "difference"}, otherwise \code{scores_pre - scores_post} is given.
If \code{minimize == FALSE}, then the order is flipped, insuring that "higher value" means "more important".
}
