% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{compute_score_relation}
\alias{compute_score_relation}
\title{Scoring utility}
\usage{
compute_score_relation(
  scores_pre,
  scores_post,
  relation = "difference",
  minimize = TRUE
)
}
\arguments{
\item{scores_pre}{(numeric()) Score before change.}

\item{scores_post}{(numeric()) Score after change.}

\item{relation}{(character(1), \code{"difference"}) Either \code{"difference"} or \code{"ratio"}.}

\item{minimize}{(logical(1), \code{TRUE}) Whether the score needs to be minimized (e.g. RMSE) or
maximized (e.g. AUC).}
}
\value{
A numeric vector of the same length as \code{scores_pre} and \code{scores_post}
}
\description{
Computes the relation of score before a change (e.g. permutation, LOCO, ...) and after.
}
\details{
If \code{minimize == TRUE}, then \code{scores_post - scores_pre} is computed for \code{relation == "difference"},
otherwise \code{scores_pre - scores_post} is given.
}
\examples{

pre = rnorm(10)
post = pre + runif(10)

compute_score_relation(pre, post)
compute_score_relation(pre, post, "ratio")
compute_score_relation(pre, post, minimize = FALSE)
}
