% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureImportanceLearner.R
\name{FeatureImportanceLearner}
\alias{FeatureImportanceLearner}
\title{Feature Importance Learner Class}
\description{
Feature Importance Learner Class

Feature Importance Learner Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{label}}{(\code{\link{character(1)}}) Method label}

\item{\code{task}}{(\code{\link[mlr3:Task]{mlr3::Task}})}

\item{\code{learner}}{(\code{\link[mlr3:Learner]{mlr3::Learner}})}

\item{\code{measure}}{(\code{\link[mlr3:Measure]{mlr3::Measure}})}

\item{\code{resampling}}{(\code{\link[mlr3:Resampling]{mlr3::Resampling}})}

\item{\code{resample_result}}{(\code{\link[mlr3:ResampleResult]{mlr3::ResampleResult}})}

\item{\code{features}}{(\code{\link[=list]{list()}})}

\item{\code{param_set}}{(\code{\link[paradox:ps]{paradox::ps()}})}

\item{\code{importance}}{(\code{\link[=data.table]{data.table()}}) Aggregated importance scores}

\item{\code{scores}}{(\code{\link[=data.table]{data.table()}}) Individual performance scores used to compute \verb{$importance}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FeatureImportanceLearner-new}{\code{FeatureImportanceLearner$new()}}
\item \href{#method-FeatureImportanceLearner-combine}{\code{FeatureImportanceLearner$combine()}}
\item \href{#method-FeatureImportanceLearner-print}{\code{FeatureImportanceLearner$print()}}
\item \href{#method-FeatureImportanceLearner-clone}{\code{FeatureImportanceLearner$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureImportanceLearner-new"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureImportanceLearner-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
This is typically intended for use by derived classes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImportanceLearner$new(
  task,
  learner,
  measure,
  resampling = NULL,
  features = NULL,
  param_set = paradox::ps(),
  label
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task, learner, measure, resampling, features, param_set, label}}{Used to set fields}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureImportanceLearner-combine"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureImportanceLearner-combine}{}}}
\subsection{Method \code{combine()}}{
Combine two \code{FeatureImportanceLearner} objects with computed scores.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImportanceLearner$combine(y, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y}}{(\link{FeatureImportanceLearner}) Object to combine. Must have computed scores.}

\item{\code{...}}{(any) Unused.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \link{FeatureImportanceLearner} of the same subclass as \code{x} and \code{y}.
Currently this method merges the following:
\itemize{
\item \verb{$scores} is combined, with \code{iter_rsmp} increased for \code{y}.
\item \verb{$importance} is re-computed from the combined \verb{$scores}.
\item \verb{$resample_result} is combined to a \link[mlr3:BenchmarkResult]{mlr3::BenchmarkResult}
\item \verb{$resampling} is combined into a \link[mlr3:mlr_resamplings_custom]{mlr3::ResamplingCustom}, again continuing te \code{iteration} count from \code{x} with that of \code{y}.
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureImportanceLearner-print"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureImportanceLearner-print}{}}}
\subsection{Method \code{print()}}{
Print importance scores
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImportanceLearner$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Passed to \code{print()}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureImportanceLearner-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureImportanceLearner-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImportanceLearner$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
