% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureImportanceMeasure.R
\name{FeatureImportanceMeasure}
\alias{FeatureImportanceMeasure}
\title{Feature Importance Learner Class}
\description{
Feature Importance Learner Class

Feature Importance Learner Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{label}}{(\code{character(1)}) Method label.}

\item{\code{task}}{(\link[mlr3:Task]{mlr3::Task})}

\item{\code{learner}}{(\link[mlr3:Learner]{mlr3::Learner})}

\item{\code{measure}}{(\link[mlr3:Measure]{mlr3::Measure})}

\item{\code{resampling}}{(\link[mlr3:Resampling]{mlr3::Resampling})}

\item{\code{resample_result}}{(\link[mlr3:ResampleResult]{mlr3::ResampleResult})}

\item{\code{features}}{(\code{character})}

\item{\code{param_set}}{(\code{\link[paradox:ps]{paradox::ps()}})}

\item{\code{importance}}{(\link[data.table:data.table]{data.table}) Aggregated importance scores}

\item{\code{scores}}{(\link[data.table:data.table]{data.table}) Individual performance scores used to compute \verb{$importance} per resampling iteration and permutation iteration.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FeatureImportanceMeasure-new}{\code{FeatureImportanceMeasure$new()}}
\item \href{#method-FeatureImportanceMeasure-compute}{\code{FeatureImportanceMeasure$compute()}}
\item \href{#method-FeatureImportanceMeasure-combine}{\code{FeatureImportanceMeasure$combine()}}
\item \href{#method-FeatureImportanceMeasure-print}{\code{FeatureImportanceMeasure$print()}}
\item \href{#method-FeatureImportanceMeasure-clone}{\code{FeatureImportanceMeasure$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureImportanceMeasure-new"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureImportanceMeasure-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
This is typically intended for use by derived classes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImportanceMeasure$new(
  task,
  learner,
  measure,
  resampling = NULL,
  features = NULL,
  param_set = paradox::ps(),
  label
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task, learner, measure, resampling, features, param_set, label}}{Used to set fields}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureImportanceMeasure-compute"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureImportanceMeasure-compute}{}}}
\subsection{Method \code{compute()}}{
Compute feature importance scores
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImportanceMeasure$compute(
  relation = c("difference", "ratio"),
  store_backends = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{relation}}{(\code{character(1): "difference"}) How to relate perturbed scores to originals ("difference" or "ratio")}

\item{\code{store_backends}}{(\code{logical(1): TRUE}) Whether to store backends.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureImportanceMeasure-combine"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureImportanceMeasure-combine}{}}}
\subsection{Method \code{combine()}}{
Combine two \code{FeatureImportanceMeasure} objects with computed scores.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImportanceMeasure$combine(y, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y}}{(\link{FeatureImportanceMeasure}) Object to combine. Must have computed scores.}

\item{\code{...}}{(any) Unused.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \link{FeatureImportanceMeasure} of the same subclass as \code{x} and \code{y}.
Currently this method merges the following:
\itemize{
\item \verb{$scores} is combined, with \code{iter_rsmp} increased for \code{y}.
\item \verb{$importance} is re-computed from the combined \verb{$scores}.
\item \verb{$resample_result} is combined to a \link[mlr3:BenchmarkResult]{mlr3::BenchmarkResult}
\item \verb{$resampling} is combined into a \link[mlr3:mlr_resamplings_custom]{mlr3::ResamplingCustom}, again continuing te \code{iteration} count from \code{x} with that of \code{y}.
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureImportanceMeasure-print"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureImportanceMeasure-print}{}}}
\subsection{Method \code{print()}}{
Print importance scores
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImportanceMeasure$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Passed to \code{print()}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureImportanceMeasure-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureImportanceMeasure-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImportanceMeasure$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
