% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSampler.R
\name{FeatureSampler}
\alias{FeatureSampler}
\title{Feature Sampler Class}
\description{
Base class for implementing different sampling strategies
for feature importance methods like PFI and CFI
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{task}}{(\link[mlr3:Task]{mlr3::Task}) Original task.}

\item{\code{label}}{(\code{character(1)}) Name of the sampler.}

\item{\code{param_set}}{(\link[paradox:ParamSet]{paradox::ParamSet}) Parameter set for the sampler.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FeatureSampler-new}{\code{FeatureSampler$new()}}
\item \href{#method-FeatureSampler-sample}{\code{FeatureSampler$sample()}}
\item \href{#method-FeatureSampler-clone}{\code{FeatureSampler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureSampler-new"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureSampler-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the FeatureSampler class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSampler$new(task)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{(\link[mlr3:Task]{mlr3::Task}) Task to sample from}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureSampler-sample"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureSampler-sample}{}}}
\subsection{Method \code{sample()}}{
Sample values for feature(s)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSampler$sample(feature, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature}}{(\code{character}) Feature name(s) to sample (can be single or multiple)}

\item{\code{data}}{(\code{\link[data.table:data.table]{data.table}} ) Data to use for sampling context}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Modified copy of the input data with the feature(s) sampled
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureSampler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureSampler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureSampler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
