% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op-null-default.R
\name{op-null-default}
\alias{op-null-default}
\alias{\%||\%}
\title{Default value for \code{NULL}}
\usage{
x \%||\% y
}
\arguments{
\item{x, y}{If \code{x} is NULL or length 0, will return \code{y}; otherwise returns \code{x}.}
}
\description{
A backport of \verb{\%||\%} available in R versions from 4.4.0.
}
\examples{
1 \%||\% 2
NULL \%||\% 2
}
