% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibentries.R
\name{print_bib}
\alias{print_bib}
\title{Print an Rd-formatted bib entry}
\usage{
print_bib(...)
}
\arguments{
\item{...}{(\code{character}) One or more quoted names of \code{bibentries} to print.}
}
\description{
Print an Rd-formatted bib entry
}
\keyword{internal}
