% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSampler.R
\name{RelativeSampler}
\alias{RelativeSampler}
\title{Relative Feature Sampler}
\description{
Implements sampling for RFI, where the feature is sampled
conditionally on a user-specified set of features
}
\section{Super classes}{
\code{\link[xplainfi:FeatureSampler]{xplainfi::FeatureSampler}} -> \code{\link[xplainfi:ConditionalSampler]{xplainfi::ConditionalSampler}} -> \code{RelativeSampler}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RelativeSampler-sample}{\code{RelativeSampler$sample()}}
\item \href{#method-RelativeSampler-clone}{\code{RelativeSampler$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="xplainfi" data-topic="ConditionalSampler" data-id="initialize"><a href='../../xplainfi/html/ConditionalSampler.html#method-ConditionalSampler-initialize'><code>xplainfi::ConditionalSampler$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RelativeSampler-sample"></a>}}
\if{latex}{\out{\hypertarget{method-RelativeSampler-sample}{}}}
\subsection{Method \code{sample()}}{
Sample values for a feature conditionally on a specified set of features
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelativeSampler$sample(feature, data, conditioning_set)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature}}{(\link{character(1)}) Feature name to sample}

\item{\code{data}}{(\link{data.table}) Data containing conditioning features}

\item{\code{conditioning_set}}{(\link{character}) Specific set of features to condition on}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Modified copy of the input data with the feature sampled conditionally
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RelativeSampler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RelativeSampler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelativeSampler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
