% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pkg.R
\name{require_package}
\alias{require_package}
\title{Package dependency helper}
\usage{
require_package(pkg, from = "cran")
}
\arguments{
\item{pkg}{(\code{character(1)}) Package name to check.}

\item{from}{(\code{character(1)}: \code{"cran"}) Package source. If not \code{"cran"},
should be the URL to the package remote, e.g. a GitHub URL.}
}
\description{
Error if the requried package is not installed and show an informative
message to the user about how to install it, defaulting to installation
from CRAN via \code{install.packages()} or some other source, e.g. a GitHub URL.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
require_package("arf")
require_package("seqknockoff", from = "https://github.com/kormama1/seqknockoff")
\dontshow{\}) # examplesIf}
}
