% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-params.R
\name{resolve_param}
\alias{resolve_param}
\title{Hierarchical parameter resolution helper}
\usage{
resolve_param(arg, stored, default)
}
\arguments{
\item{arg, stored, default}{Scalar parameter values evlauted in that order}
}
\value{
First non-\code{NULL} parameter
}
\description{
Used to retrieve the highest-priority parameter value.
}
\details{
Priority is \code{arg} > \code{stored} > \code{default}.
}
\examples{
resolve_param(arg = "foo", stored = "bar", default = "baz")
resolve_param(arg = NULL, stored = "bar", default = "baz")
resolve_param(arg = NULL, stored = NULL, default = "baz")
}
\keyword{internal}
