% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sim.R
\name{sim_dgp_ewald}
\alias{sim_dgp_ewald}
\title{Simulate data as in Ewald et al. (2024)}
\usage{
sim_dgp_ewald(n = 500)
}
\arguments{
\item{n}{(\code{integer(1)}) Number of samples to create.}
}
\value{
A regression task (\link[mlr3:TaskRegr]{mlr3::TaskRegr}) with \link[data.table:data.table]{data.table} backend.
}
\description{
– X1, X3 and X5 are independent and standard normal: Xj ∼ N (0, 1),
– X2 is a noisy copy of X1: X2 := X1 + eps_2, eps_2 ∼ N (0, 0.001),
– X4 is a (more) noisy copy of X3: X4 := X3 + eps_4, eps_4 ∼ N (0, 0.1),
– Y depends on X4 and X5 via linear effects and a bivariate interaction: Y := X4 + X5 + X4 ∗ X5 + eps_Y , eps_Y ∼ N (0, 0.1).
}
\examples{
sim_dgp_ewald(100)

}
