% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sim.R
\name{sim_dgp_example}
\alias{sim_dgp_example}
\title{Example DGP for illustration}
\usage{
sim_dgp_example(n = 100L)
}
\arguments{
\item{n}{(\code{integer(1)}: 100L) Number of samples to create.}
}
\value{
A regression task (\link[mlr3:TaskRegr]{mlr3::TaskRegr}) with \link[data.table:data.table]{data.table} backend.
}
\description{
\itemize{
\item x1: random uniform, direct effect on y
\item xm, xme: random uniform, xme is the mediated exposure, xm is the mediator, no direct effect of xme on y
\item xi1,2: random uniform, interaction effect only
\item z: random uniform, xz is noisy version of z, only z has effect on y
\item xc1,2,3: correlated multivariate normal, additive effects on y
\item xbin: random bernoulli with prob 1/3
\item znoise: random uniform and random normal noise terms, no effect on y
}
}
\details{
\eqn{y = x1 + x2 + xi1 * xi2 + z + xc1 + xc2 + xc3 + xbin + \epsilon}
}
\examples{
sim_dgp_example(100)
}
