% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WVIM.R
\name{WVIM}
\alias{WVIM}
\title{Williamson's Variable Importance Measure (WVIM)}
\description{
Base class generalizazing refit-based variable importance measures.
Default corresponds to leaving out each feature \code{iters_refit} times, which
corresponds to LOCO (Leave One Covariate Out).
}
\keyword{internal}
\section{Super class}{
\code{\link[xplainfi:FeatureImportanceMethod]{xplainfi::FeatureImportanceMethod}} -> \code{WVIM}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{direction}}{(\code{character(1)}) Either "leave-out" or "leave-in".}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WVIM-new}{\code{WVIM$new()}}
\item \href{#method-WVIM-compute}{\code{WVIM$compute()}}
\item \href{#method-WVIM-clone}{\code{WVIM$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="xplainfi" data-topic="FeatureImportanceMethod" data-id="importance"><a href='../../xplainfi/html/FeatureImportanceMethod.html#method-FeatureImportanceMethod-importance'><code>xplainfi::FeatureImportanceMethod$importance()</code></a></span></li>
<li><span class="pkg-link" data-pkg="xplainfi" data-topic="FeatureImportanceMethod" data-id="obs_scores"><a href='../../xplainfi/html/FeatureImportanceMethod.html#method-FeatureImportanceMethod-obs_scores'><code>xplainfi::FeatureImportanceMethod$obs_scores()</code></a></span></li>
<li><span class="pkg-link" data-pkg="xplainfi" data-topic="FeatureImportanceMethod" data-id="print"><a href='../../xplainfi/html/FeatureImportanceMethod.html#method-FeatureImportanceMethod-print'><code>xplainfi::FeatureImportanceMethod$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="xplainfi" data-topic="FeatureImportanceMethod" data-id="reset"><a href='../../xplainfi/html/FeatureImportanceMethod.html#method-FeatureImportanceMethod-reset'><code>xplainfi::FeatureImportanceMethod$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="xplainfi" data-topic="FeatureImportanceMethod" data-id="scores"><a href='../../xplainfi/html/FeatureImportanceMethod.html#method-FeatureImportanceMethod-scores'><code>xplainfi::FeatureImportanceMethod$scores()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WVIM-new"></a>}}
\if{latex}{\out{\hypertarget{method-WVIM-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WVIM$new(
  task,
  learner,
  measure,
  resampling = NULL,
  features = NULL,
  direction = c("leave-out", "leave_in"),
  iters_refit = 1L
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task, learner, measure, resampling, features}}{Passed to \code{FeatureImportanceMethod} for construction.}

\item{\code{direction}}{(\code{character(1)}) Either "leave-out" or "leave-in".}

\item{\code{iters_refit}}{(\code{integer(1)}) Number of refit iterations per resampling iteration.}

\item{\code{label}}{(\code{character(1)}) Method label.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WVIM-compute"></a>}}
\if{latex}{\out{\hypertarget{method-WVIM-compute}{}}}
\subsection{Method \code{compute()}}{
Computes leave-out or leave-in feature importance.
\code{wvim_design_matrix(task$feature_names, "leave-out")} corresponds to LOCO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WVIM$compute(store_backends = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{store_backends}}{(\code{logical(1)}) Passed to \link[mlr3:resample]{mlr3::resample} to store
backends in resample result.
Required for some measures, but may increase memory footprint.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WVIM-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WVIM-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WVIM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
