% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplainfi-package.R
\docType{package}
\name{xplainfi-package}
\alias{xplainfi}
\alias{xplainfi-package}
\title{xplainfi: Feature Importance Methods for Model Interpretability}
\description{
Provides a consistent interface for common feature importance methods, such as permutation feature importance, 'LOCO', and 'SAGE'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jemus42.github.io/xplainfi/}
  \item \url{https://github.com/jemus42/xplainfi}
  \item Report bugs at \url{https://github.com/jemus42/xplainfi/issues}
}

}
\author{
\strong{Maintainer}: Lukas Burk \email{github@quantenbrot.de} (\href{https://orcid.org/0000-0001-7528-3795}{ORCID})

}
\keyword{internal}
