% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{c.FeatureImportanceMeasure}
\alias{c.FeatureImportanceMeasure}
\title{Combine two FeatureImportanceMeasure objects}
\usage{
\method{c}{FeatureImportanceMeasure}(x, y, ...)
}
\arguments{
\item{x, y}{\verb{([FeatureImportanceMeasure])} Objects to combine. Must have computed scores.}

\item{...}{(any) Ignored.}
}
\value{
New object of the same subclass as \code{x} and \code{y}.
}
\description{
Combine two FeatureImportanceMeasure objects
}
\examples{
if (FALSE) {

}
}
