% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerturbationImportance.R
\name{CFI}
\alias{CFI}
\title{Conditional Feature Importance}
\description{
Implementation of CFI using modular sampling approach
}
\section{Super classes}{
\code{\link[xplainfi:FeatureImportanceMeasure]{xplainfi::FeatureImportanceMeasure}} -> \code{\link[xplainfi:PerturbationImportance]{xplainfi::PerturbationImportance}} -> \code{CFI}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CFI-new}{\code{CFI$new()}}
\item \href{#method-CFI-compute}{\code{CFI$compute()}}
\item \href{#method-CFI-clone}{\code{CFI$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="xplainfi" data-topic="FeatureImportanceMeasure" data-id="combine"><a href='../../xplainfi/html/FeatureImportanceMeasure.html#method-FeatureImportanceMeasure-combine'><code>xplainfi::FeatureImportanceMeasure$combine()</code></a></span></li>
<li><span class="pkg-link" data-pkg="xplainfi" data-topic="PerturbationImportance" data-id="print"><a href='../../xplainfi/html/PerturbationImportance.html#method-PerturbationImportance-print'><code>xplainfi::PerturbationImportance$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CFI-new"></a>}}
\if{latex}{\out{\hypertarget{method-CFI-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the CFI class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CFI$new(
  task,
  learner,
  measure,
  resampling = NULL,
  features = NULL,
  iters_perm = 1L,
  sampler = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task, learner, measure, resampling, features}}{Passed to \code{PerturbationImportance}.}

\item{\code{iters_perm}}{(integer(1)) Number of sampling iterations.}

\item{\code{sampler}}{(\link{ConditionalSampler}) Optional custom sampler. Defaults to \code{ARFSampler}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CFI-compute"></a>}}
\if{latex}{\out{\hypertarget{method-CFI-compute}{}}}
\subsection{Method \code{compute()}}{
Compute CFI scores
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CFI$compute(relation = c("difference", "ratio"), store_backends = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{relation}}{(character(1)) How to relate perturbed scores to originals}

\item{\code{store_backends}}{(logical(1)) Whether to store backends}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CFI-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CFI-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CFI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
