% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerturbationImportance.R
\name{CFI}
\alias{CFI}
\title{Conditional Feature Importance}
\description{
Implementation of CFI using modular sampling approach
}
\examples{
\dontshow{if (requireNamespace("ranger", quietly = TRUE) && requireNamespace("mlr3learners", quietly = TRUE) && requireNamespace("arf", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(mlr3)
task = tgen("2dnormals")$generate(n = 100)
cfi = CFI$new(
  task = task,
  learner = lrn("classif.ranger", num.trees = 50, predict_type = "prob"),
  measure = msr("classif.ce")
)
cfi$compute()
\dontshow{\}) # examplesIf}
}
\references{
Blesch, Kristin, Koenen, Niklas, Kapar, Jan, Golchian, Pegah, Burk, Lukas, Loecher, Markus, Wright, N. M (2025).
\dQuote{Conditional Feature Importance with Generative Modeling Using Adversarial Random Forests.}
\emph{Proceedings of the AAAI Conference on Artificial Intelligence}, \bold{39}(15), 15596--15604.
\doi{10.1609/aaai.v39i15.33712}.
}
\section{Super classes}{
\code{\link[xplainfi:FeatureImportanceMeasure]{xplainfi::FeatureImportanceMeasure}} -> \code{\link[xplainfi:PerturbationImportance]{xplainfi::PerturbationImportance}} -> \code{CFI}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CFI-new}{\code{CFI$new()}}
\item \href{#method-CFI-compute}{\code{CFI$compute()}}
\item \href{#method-CFI-clone}{\code{CFI$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="xplainfi" data-topic="FeatureImportanceMeasure" data-id="combine"><a href='../../xplainfi/html/FeatureImportanceMeasure.html#method-FeatureImportanceMeasure-combine'><code>xplainfi::FeatureImportanceMeasure$combine()</code></a></span></li>
<li><span class="pkg-link" data-pkg="xplainfi" data-topic="FeatureImportanceMeasure" data-id="print"><a href='../../xplainfi/html/FeatureImportanceMeasure.html#method-FeatureImportanceMeasure-print'><code>xplainfi::FeatureImportanceMeasure$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CFI-new"></a>}}
\if{latex}{\out{\hypertarget{method-CFI-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the CFI class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CFI$new(
  task,
  learner,
  measure,
  resampling = NULL,
  features = NULL,
  iters_perm = 1L,
  sampler = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task, learner, measure, resampling, features}}{Passed to \code{PerturbationImportance}.}

\item{\code{iters_perm}}{(integer(1)) Number of sampling iterations.}

\item{\code{sampler}}{(\link{ConditionalSampler}) Optional custom sampler. Defaults to instantiationg \code{ARFSampler} internally with default parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CFI-compute"></a>}}
\if{latex}{\out{\hypertarget{method-CFI-compute}{}}}
\subsection{Method \code{compute()}}{
Compute CFI scores
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CFI$compute(relation = c("difference", "ratio"), store_backends = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{relation}}{(character(1)) How to relate perturbed scores to originals}

\item{\code{store_backends}}{(logical(1)) Whether to store backends}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CFI-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CFI-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CFI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
