% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_score.R
\name{compute_score}
\alias{compute_score}
\title{Scoring utility}
\usage{
compute_score(
  scores_pre,
  scores_post,
  relation = c("difference", "ratio"),
  minimize = TRUE
)
}
\arguments{
\item{scores_pre}{(numeric()) Score before change.}

\item{scores_post}{(numeric()) Score after change.}

\item{relation}{(character(1), \code{"difference"}) Either \code{"difference"} or \code{"ratio"}.
If \code{"difference"}, then \code{scores_post - scores_pre} is computed, otherwise \code{scores_post / scores_pre}.}

\item{minimize}{(logical(1), \code{TRUE}) Whether the score needs to be minimized (e.g. RMSE) or
maximized (e.g. AUC).}
}
\value{
A numeric vector of the same length as \code{scores_pre} and \code{scores_post}
}
\description{
Computes the relation of score before a change (e.g. \link{PFI}, \link{LOCO}, ...) and after.
}
\details{
If \code{minimize == TRUE}, then \code{scores_post - scores_pre} is computed for
\code{relation == "difference"}, otherwise \code{scores_pre - scores_post} is given.
If \code{minimize == FALSE}, then \code{scores_pre - scores_post} is computed.
}
\examples{

pre = rnorm(10)
post = pre + runif(10)

compute_score(pre, post)
compute_score(pre, post, "ratio")
compute_score(pre, post, minimize = FALSE)
}
