% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSampler.R
\name{ConditionalSampler}
\alias{ConditionalSampler}
\title{Conditional Feature Sampler}
\description{
Base class for conditional sampling methods where features
are sampled conditionally on other features. This is an abstract class
that should be extended by concrete implementations.
}
\section{Super class}{
\code{\link[xplainfi:FeatureSampler]{xplainfi::FeatureSampler}} -> \code{ConditionalSampler}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConditionalSampler-new}{\code{ConditionalSampler$new()}}
\item \href{#method-ConditionalSampler-sample}{\code{ConditionalSampler$sample()}}
\item \href{#method-ConditionalSampler-clone}{\code{ConditionalSampler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConditionalSampler-new"></a>}}
\if{latex}{\out{\hypertarget{method-ConditionalSampler-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the ConditionalSampler class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionalSampler$new(task)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{(\link[mlr3:Task]{mlr3::Task}) Task to sample from}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConditionalSampler-sample"></a>}}
\if{latex}{\out{\hypertarget{method-ConditionalSampler-sample}{}}}
\subsection{Method \code{sample()}}{
Sample values for a feature conditionally on all other features
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionalSampler$sample(feature, data, conditioning_features = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature}}{(\link{character(1)}) Feature name to sample}

\item{\code{data}}{(\link{data.table}) Data containing conditioning features}

\item{\code{conditioning_features}}{(\link{character}) Features to condition on (default: all other features)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Modified copy of the input data with the feature sampled conditionally
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConditionalSampler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ConditionalSampler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionalSampler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
