% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureImportanceLearner.R
\name{FeatureImportanceLearner}
\alias{FeatureImportanceLearner}
\title{Feature Importance Learner Class}
\description{
Feature Importance Learner Class

Feature Importance Learner Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{label}}{(character(1)) Method label}

\item{\code{task}}{(mlr3::Task)}

\item{\code{learner}}{(mlr3::Learner)}

\item{\code{measure}}{(mlr3::Measure)}

\item{\code{resampling}}{(mlr3::Resampling)}

\item{\code{resample_result}}{(mlr3::ResampleResult)}

\item{\code{features}}{(list())}

\item{\code{param_set}}{(paradox::ps())}

\item{\code{importance}}{(numeric())}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FeatureImportanceLearner-new}{\code{FeatureImportanceLearner$new()}}
\item \href{#method-FeatureImportanceLearner-print}{\code{FeatureImportanceLearner$print()}}
\item \href{#method-FeatureImportanceLearner-clone}{\code{FeatureImportanceLearner$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureImportanceLearner-new"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureImportanceLearner-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
This is typically intended for use by derived classes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImportanceLearner$new(
  task,
  learner,
  measure,
  resampling = NULL,
  features = NULL,
  param_set = paradox::ps(),
  label
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task, learner, measure, resampling, features, param_set, label}}{Used to set fields}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureImportanceLearner-print"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureImportanceLearner-print}{}}}
\subsection{Method \code{print()}}{
Print importance scores
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImportanceLearner$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Passed to \code{print()}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureImportanceLearner-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureImportanceLearner-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureImportanceLearner$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
