% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAGE.R
\name{MarginalSAGE}
\alias{MarginalSAGE}
\title{Marginal SAGE}
\description{
SAGE with marginal sampling (features are marginalized independently).
This is the standard SAGE implementation.
}
\examples{
\dontshow{if (requireNamespace("ranger", quietly = TRUE) && requireNamespace("mlr3learners", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(mlr3)
task = tgen("friedman1")$generate(n = 100)
sage = MarginalSAGE$new(
  task = task,
  learner = lrn("regr.ranger", num.trees = 50),
  measure = msr("regr.mse"),
  n_permutations = 3L
)
sage$compute()
\dontshow{\}) # examplesIf}
}
\section{Super classes}{
\code{\link[xplainfi:FeatureImportanceMethod]{xplainfi::FeatureImportanceMethod}} -> \code{\link[xplainfi:SAGE]{xplainfi::SAGE}} -> \code{MarginalSAGE}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MarginalSAGE-new}{\code{MarginalSAGE$new()}}
\item \href{#method-MarginalSAGE-clone}{\code{MarginalSAGE$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="xplainfi" data-topic="FeatureImportanceMethod" data-id="combine"><a href='../../xplainfi/html/FeatureImportanceMethod.html#method-FeatureImportanceMethod-combine'><code>xplainfi::FeatureImportanceMethod$combine()</code></a></span></li>
<li><span class="pkg-link" data-pkg="xplainfi" data-topic="FeatureImportanceMethod" data-id="print"><a href='../../xplainfi/html/FeatureImportanceMethod.html#method-FeatureImportanceMethod-print'><code>xplainfi::FeatureImportanceMethod$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="xplainfi" data-topic="SAGE" data-id="compute"><a href='../../xplainfi/html/SAGE.html#method-SAGE-compute'><code>xplainfi::SAGE$compute()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarginalSAGE-new"></a>}}
\if{latex}{\out{\hypertarget{method-MarginalSAGE-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the MarginalSAGE class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginalSAGE$new(
  task,
  learner,
  measure,
  resampling = NULL,
  features = NULL,
  n_permutations = 10L,
  reference_data = NULL,
  max_reference_size = 100L
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task, learner, measure, resampling, features}}{Passed to \link{SAGE}.}

\item{\code{n_permutations}}{(integer(1)) Number of permutations to sample.}

\item{\code{reference_data}}{(data.table) Optional reference dataset.}

\item{\code{max_reference_size}}{(integer(1)) Maximum size of reference dataset.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarginalSAGE-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MarginalSAGE-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginalSAGE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
