% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSampler-Marginal.R
\name{MarginalSampler}
\alias{MarginalSampler}
\title{Marginal Feature Sampler}
\description{
Implements marginal sampling for PFI, where the feature of interest
is sampled independently of other features
}
\examples{
library(mlr3)
task = tgen("2dnormals")$generate(n = 100)
sampler = MarginalSampler$new(task)
# Sample using row_ids from stored task
sampled_data = sampler$sample("x1")
# Or use external data
data = task$data()
sampled_data_ext = sampler$sample_newdata("x1", newdata = data)
}
\section{Super class}{
\code{\link[xplainfi:FeatureSampler]{xplainfi::FeatureSampler}} -> \code{MarginalSampler}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MarginalSampler-new}{\code{MarginalSampler$new()}}
\item \href{#method-MarginalSampler-sample}{\code{MarginalSampler$sample()}}
\item \href{#method-MarginalSampler-sample_newdata}{\code{MarginalSampler$sample_newdata()}}
\item \href{#method-MarginalSampler-clone}{\code{MarginalSampler$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="xplainfi" data-topic="FeatureSampler" data-id="print"><a href='../../xplainfi/html/FeatureSampler.html#method-FeatureSampler-print'><code>xplainfi::FeatureSampler$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarginalSampler-new"></a>}}
\if{latex}{\out{\hypertarget{method-MarginalSampler-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the MarginalSampler class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginalSampler$new(task)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{(\link[mlr3:Task]{mlr3::Task}) Task to sample from}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarginalSampler-sample"></a>}}
\if{latex}{\out{\hypertarget{method-MarginalSampler-sample}{}}}
\subsection{Method \code{sample()}}{
Sample from stored task by permutation (marginal distribution).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginalSampler$sample(feature, row_ids = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature}}{(\code{character}) Feature(s) to sample.}

\item{\code{row_ids}}{(\code{integer()} | \code{NULL}) Row IDs to use. If \code{NULL}, uses all rows.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Modified copy with permuted feature(s).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarginalSampler-sample_newdata"></a>}}
\if{latex}{\out{\hypertarget{method-MarginalSampler-sample_newdata}{}}}
\subsection{Method \code{sample_newdata()}}{
Sample from external data by permutation. See \verb{$sample()} for details.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginalSampler$sample_newdata(feature, newdata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature}}{(\code{character}) Feature(s) to sample.}

\item{\code{newdata}}{(\code{\link[data.table:data.table]{data.table}}) External data to use.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Modified copy with permuted feature(s).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarginalSampler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MarginalSampler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginalSampler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
