% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSampler.R
\name{MarginalSampler}
\alias{MarginalSampler}
\title{Marginal Feature Sampler}
\description{
Implements marginal sampling for PFI, where the feature of interest
is sampled independently of other features
}
\examples{
library(mlr3)
task = tgen("2dnormals")$generate(n = 100)
sampler = MarginalSampler$new(task)
data = task$data()
sampled_data = sampler$sample("x1", data)
}
\section{Super class}{
\code{\link[xplainfi:FeatureSampler]{xplainfi::FeatureSampler}} -> \code{MarginalSampler}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MarginalSampler-new}{\code{MarginalSampler$new()}}
\item \href{#method-MarginalSampler-sample}{\code{MarginalSampler$sample()}}
\item \href{#method-MarginalSampler-print}{\code{MarginalSampler$print()}}
\item \href{#method-MarginalSampler-clone}{\code{MarginalSampler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarginalSampler-new"></a>}}
\if{latex}{\out{\hypertarget{method-MarginalSampler-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the MarginalSampler class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginalSampler$new(task)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{(\link[mlr3:Task]{mlr3::Task}) Task to sample from}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarginalSampler-sample"></a>}}
\if{latex}{\out{\hypertarget{method-MarginalSampler-sample}{}}}
\subsection{Method \code{sample()}}{
Sample values for feature(s) by permutation (marginal distribution)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginalSampler$sample(feature, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature}}{(\code{character}) Feature name(s) to sample (can be single or multiple)}

\item{\code{data}}{(\code{\link[data.table:data.table]{data.table}} ) Data to permute the feature(s) in}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Modified copy of the input data with the feature(s) permuted
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarginalSampler-print"></a>}}
\if{latex}{\out{\hypertarget{method-MarginalSampler-print}{}}}
\subsection{Method \code{print()}}{
Print sampler
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginalSampler$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Passed to \code{print()}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarginalSampler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MarginalSampler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginalSampler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
