% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFI.R
\name{PFI}
\alias{PFI}
\title{PFI Class}
\description{
PFI Class

PFI Class
}
\details{
Calculates Permutation Feature Importance (PFI) scores.
}
\examples{
\dontshow{if (requireNamespace("rpart")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

library(mlr3)

pfi = PFI$new(
  task = tsk("zoo"),
  learner = lrn("classif.rpart"),
  measure = msr("classif.ce")
)

pfi$compute()
\dontshow{\}) # examplesIf}
}
\section{Super class}{
\code{\link[xplainfi:FeatureImportanceLearner]{xplainfi::FeatureImportanceLearner}} -> \code{PFI}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PFI-new}{\code{PFI$new()}}
\item \href{#method-PFI-compute}{\code{PFI$compute()}}
\item \href{#method-PFI-clone}{\code{PFI$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="xplainfi" data-topic="FeatureImportanceLearner" data-id="combine"><a href='../../xplainfi/html/FeatureImportanceLearner.html#method-FeatureImportanceLearner-combine'><code>xplainfi::FeatureImportanceLearner$combine()</code></a></span></li>
<li><span class="pkg-link" data-pkg="xplainfi" data-topic="FeatureImportanceLearner" data-id="print"><a href='../../xplainfi/html/FeatureImportanceLearner.html#method-FeatureImportanceLearner-print'><code>xplainfi::FeatureImportanceLearner$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PFI-new"></a>}}
\if{latex}{\out{\hypertarget{method-PFI-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PFI$new(
  task,
  learner,
  measure,
  resampling = NULL,
  features = NULL,
  iters_perm = 1L
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task, learner, measure, resampling, features}}{Passed to \code{FeatureImportanceLearner} for construction.}

\item{\code{iters_perm}}{\code{(integer(1): 1L)} Number of permutations to compute for each feature.
Permutations are repeated within each resampling fold.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PFI-compute"></a>}}
\if{latex}{\out{\hypertarget{method-PFI-compute}{}}}
\subsection{Method \code{compute()}}{
A short description...
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PFI$compute(relation = c("difference", "ratio"), store_backends = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{relation}}{(character(1)) Calculate \code{"difference"} (default) or \code{"ratio"} of
original scores and scores after permutation}

\item{\code{store_backends}}{(logical(1): \code{TRUE}) Passed to \code{\link[=resample]{resample()}} to store
backends in resample result.
Required for some measures, but may increase memory footprint.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PFI-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PFI-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PFI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
