% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerturbationImportance.R
\name{PFI}
\alias{PFI}
\title{Permutation Feature Importance}
\description{
Implementation of PFI using modular sampling approach
}
\examples{
\dontshow{if (requireNamespace("ranger", quietly = TRUE) && requireNamespace("mlr3learners", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(mlr3)
task = tgen("2dnormals")$generate(n = 100)
pfi = PFI$new(
  task = task,
  learner = lrn("classif.ranger", num.trees = 50, predict_type = "prob"),
  measure = msr("classif.ce")
)
pfi$compute()
\dontshow{\}) # examplesIf}
}
\section{Super classes}{
\code{\link[xplainfi:FeatureImportanceMeasure]{xplainfi::FeatureImportanceMeasure}} -> \code{\link[xplainfi:PerturbationImportance]{xplainfi::PerturbationImportance}} -> \code{PFI}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PFI-new}{\code{PFI$new()}}
\item \href{#method-PFI-compute}{\code{PFI$compute()}}
\item \href{#method-PFI-clone}{\code{PFI$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="xplainfi" data-topic="FeatureImportanceMeasure" data-id="combine"><a href='../../xplainfi/html/FeatureImportanceMeasure.html#method-FeatureImportanceMeasure-combine'><code>xplainfi::FeatureImportanceMeasure$combine()</code></a></span></li>
<li><span class="pkg-link" data-pkg="xplainfi" data-topic="FeatureImportanceMeasure" data-id="print"><a href='../../xplainfi/html/FeatureImportanceMeasure.html#method-FeatureImportanceMeasure-print'><code>xplainfi::FeatureImportanceMeasure$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PFI-new"></a>}}
\if{latex}{\out{\hypertarget{method-PFI-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the PFI class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PFI$new(
  task,
  learner,
  measure,
  resampling = NULL,
  features = NULL,
  iters_perm = 1L
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task, learner, measure, resampling, features}}{Passed to PerturbationImportance}

\item{\code{iters_perm}}{(integer(1)) Number of permutation iterations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PFI-compute"></a>}}
\if{latex}{\out{\hypertarget{method-PFI-compute}{}}}
\subsection{Method \code{compute()}}{
Compute PFI scores
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PFI$compute(relation = c("difference", "ratio"), store_backends = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{relation}}{(character(1)) How to relate perturbed scores to originals}

\item{\code{store_backends}}{(logical(1)) Whether to store backends}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PFI-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PFI-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PFI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
