% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFI.R
\name{PFI}
\alias{PFI}
\title{PFI Class}
\description{
PFI Class

PFI Class
}
\details{
Calculates Permutation Feature Importance (PFI) scores.
}
\examples{
\dontshow{if (requireNamespace("mlr3learners") & requireNamespace("ranger")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

library(mlr3)

pfi = PFI$new(
  task = tsk("zoo"),
  learner = lrn("classif.rpart"),
  measure = msr("classif.ce")
)

pfi$compute()
\dontshow{\}) # examplesIf}
}
\section{Super class}{
\code{\link[xplainvi:FeatureImportanceLearner]{xplainvi::FeatureImportanceLearner}} -> \code{PFI}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PFI-new}{\code{PFI$new()}}
\item \href{#method-PFI-compute}{\code{PFI$compute()}}
\item \href{#method-PFI-clone}{\code{PFI$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="xplainvi" data-topic="FeatureImportanceLearner" data-id="print"><a href='../../xplainvi/html/FeatureImportanceLearner.html#method-FeatureImportanceLearner-print'><code>xplainvi::FeatureImportanceLearner$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PFI-new"></a>}}
\if{latex}{\out{\hypertarget{method-PFI-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PFI$new(task, learner, measure, resampling = NULL, features = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task, learner, measure, resampling, features}}{Passed to \code{FeatureImportanceLearner} for construction.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PFI-compute"></a>}}
\if{latex}{\out{\hypertarget{method-PFI-compute}{}}}
\subsection{Method \code{compute()}}{
A short description...
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PFI$compute(relation = c("difference", "ratio"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{relation}}{(character(1)) Calculate \code{"difference"} (default) or \code{"ratio"} of
original loss and loss after permutation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PFI-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PFI-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PFI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
