% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerturbationImportance.R
\name{RFI}
\alias{RFI}
\title{Relative Feature Importance}
\description{
Implementation of RFI using modular sampling approach
}
\examples{
\dontshow{if (requireNamespace("ranger", quietly = TRUE) && requireNamespace("mlr3learners", quietly = TRUE) && requireNamespace("arf", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(mlr3)
task = tgen("friedman1")$generate(n = 200)
rfi = RFI$new(
  task = task,
  learner = lrn("regr.ranger", num.trees = 50),
  measure = msr("regr.mse"),
  conditioning_set = c("important1")
)
rfi$compute()
\dontshow{\}) # examplesIf}
}
\references{
König, Gunnar, Molnar, Christoph, Bischl, Bernd, Grosse-Wentrup, Moritz (2021).
\dQuote{Relative Feature Importance.}
In \emph{2020 25th International Conference on Pattern Recognition (ICPR)}, 9318--9325.
\doi{10.1109/ICPR48806.2021.9413090}.
}
\section{Super classes}{
\code{\link[xplainfi:FeatureImportanceMethod]{xplainfi::FeatureImportanceMethod}} -> \code{\link[xplainfi:PerturbationImportance]{xplainfi::PerturbationImportance}} -> \code{RFI}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{conditioning_set}}{(\code{\link[=character]{character()}}) Features to condition on}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RFI-new}{\code{RFI$new()}}
\item \href{#method-RFI-compute}{\code{RFI$compute()}}
\item \href{#method-RFI-clone}{\code{RFI$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="xplainfi" data-topic="FeatureImportanceMethod" data-id="combine"><a href='../../xplainfi/html/FeatureImportanceMethod.html#method-FeatureImportanceMethod-combine'><code>xplainfi::FeatureImportanceMethod$combine()</code></a></span></li>
<li><span class="pkg-link" data-pkg="xplainfi" data-topic="FeatureImportanceMethod" data-id="print"><a href='../../xplainfi/html/FeatureImportanceMethod.html#method-FeatureImportanceMethod-print'><code>xplainfi::FeatureImportanceMethod$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RFI-new"></a>}}
\if{latex}{\out{\hypertarget{method-RFI-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the RFI class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RFI$new(
  task,
  learner,
  measure,
  resampling = NULL,
  features = NULL,
  conditioning_set = NULL,
  iters_perm = 1L,
  sampler = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task, learner, measure, resampling, features}}{Passed to PerturbationImportance}

\item{\code{conditioning_set}}{(\code{\link[=character]{character()}}) Set of features to condition on}

\item{\code{iters_perm}}{(integer(1)) Number of permutation iterations}

\item{\code{sampler}}{(\link{ConditionalSampler}) Optional custom sampler. Defaults to ARFSampler}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RFI-compute"></a>}}
\if{latex}{\out{\hypertarget{method-RFI-compute}{}}}
\subsection{Method \code{compute()}}{
Compute RFI scores
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RFI$compute(relation = c("difference", "ratio"), store_backends = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{relation}}{(character(1)) How to relate perturbed scores to originals}

\item{\code{store_backends}}{(logical(1)) Whether to store backends}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RFI-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RFI-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RFI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
