% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAGE.R
\name{SAGE}
\alias{SAGE}
\title{SAGE Class}
\description{
SAGE Class

SAGE Class
}
\details{
Calculates SAGE importance scores.
}
\examples{
\dontshow{if (requireNamespace("rpart")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

library(mlr3)

sage = SAGE$new(
  task = tsk("zoo"),
  learner = lrn("classif.rpart"),
  measure = msr("classif.ce")
)

sage$compute()
\dontshow{\}) # examplesIf}
}
\section{Super class}{
\code{\link[xplainfi:FeatureImportanceLearner]{xplainfi::FeatureImportanceLearner}} -> \code{SAGE}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SAGE-new}{\code{SAGE$new()}}
\item \href{#method-SAGE-compute}{\code{SAGE$compute()}}
\item \href{#method-SAGE-clone}{\code{SAGE$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="xplainfi" data-topic="FeatureImportanceLearner" data-id="combine"><a href='../../xplainfi/html/FeatureImportanceLearner.html#method-FeatureImportanceLearner-combine'><code>xplainfi::FeatureImportanceLearner$combine()</code></a></span></li>
<li><span class="pkg-link" data-pkg="xplainfi" data-topic="FeatureImportanceLearner" data-id="print"><a href='../../xplainfi/html/FeatureImportanceLearner.html#method-FeatureImportanceLearner-print'><code>xplainfi::FeatureImportanceLearner$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SAGE-new"></a>}}
\if{latex}{\out{\hypertarget{method-SAGE-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SAGE$new(task, learner, measure, resampling = NULL, features = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task, learner, measure, resampling, features}}{Passed to \code{FeatureImportanceLearner} for construction.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SAGE-compute"></a>}}
\if{latex}{\out{\hypertarget{method-SAGE-compute}{}}}
\subsection{Method \code{compute()}}{
A short description...
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SAGE$compute(relation = c("difference", "ratio"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{relation}}{(character(1)) Calculate \code{"difference"} (default) or \code{"ratio"} of
original scores and scores after permutation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SAGE-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SAGE-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SAGE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
