% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAGE.R
\name{SAGE}
\alias{SAGE}
\title{Shapley Additive Global Importance (SAGE)}
\description{
Implements SAGE (Shapley Additive Global Importance) for
feature importance based on Shapley values with marginalization.
SAGE quantifies feature importance by marginalizing out features and
computing the average loss difference.
}
\details{
SAGE uses Shapley values to fairly distribute the total prediction
performance among all features. Unlike perturbation-based methods,
SAGE marginalizes features by integrating over their marginal distribution.
This is approximated by averaging predictions over a background dataset.
}
\references{
Covert, I., Lundberg, S. M., & Lee, S. I. (2020).
Understanding global feature contributions through game-theoretic interpretations
of black-box models. arXiv preprint arXiv:2010.12012.
}
\section{Super class}{
\code{\link[xplainfi:FeatureImportanceMeasure]{xplainfi::FeatureImportanceMeasure}} -> \code{SAGE}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{n_permutations}}{(integer(1)) Number of permutations to sample.}

\item{\code{background_data}}{(data.table) Background dataset for marginalization.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SAGE-new}{\code{SAGE$new()}}
\item \href{#method-SAGE-compute}{\code{SAGE$compute()}}
\item \href{#method-SAGE-clone}{\code{SAGE$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="xplainfi" data-topic="FeatureImportanceMeasure" data-id="combine"><a href='../../xplainfi/html/FeatureImportanceMeasure.html#method-FeatureImportanceMeasure-combine'><code>xplainfi::FeatureImportanceMeasure$combine()</code></a></span></li>
<li><span class="pkg-link" data-pkg="xplainfi" data-topic="FeatureImportanceMeasure" data-id="print"><a href='../../xplainfi/html/FeatureImportanceMeasure.html#method-FeatureImportanceMeasure-print'><code>xplainfi::FeatureImportanceMeasure$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SAGE-new"></a>}}
\if{latex}{\out{\hypertarget{method-SAGE-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the SAGE class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SAGE$new(
  task,
  learner,
  measure,
  resampling = NULL,
  features = NULL,
  n_permutations = 10L,
  background_data = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task, learner, measure, resampling, features}}{Passed to FeatureImportanceMeasure.}

\item{\code{n_permutations}}{(integer(1)) Number of permutations to sample for Shapley value estimation.}

\item{\code{background_data}}{(data.table) Optional background dataset. If NULL, uses training data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SAGE-compute"></a>}}
\if{latex}{\out{\hypertarget{method-SAGE-compute}{}}}
\subsection{Method \code{compute()}}{
Compute SAGE values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SAGE$compute(store_backends = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{store_backends}}{(logical(1)) Whether to store backends.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SAGE-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SAGE-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SAGE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
