% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplainfi-package.R
\docType{package}
\name{xplainfi-package}
\alias{xplainfi}
\alias{xplainfi-package}
\title{xplainfi: Feature Importance Methods for Global Explanations}
\description{
Provides a consistent interface for common feature importance methods as described in Ewald et al. (2024) \doi{10.1007/978-3-031-63797-1_22}, including permutation feature importance ('PFI'), conditional and relative feature importance ('CFI', 'RFI'), 'leave-one-covariate-out' ('LOCO'), and 'Shapley Additive Global importancE' ('SAGE'), as well as feature sampling mechanisms to support conditional importance methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr-org.github.io/xplainfi/}
  \item \url{https://github.com/mlr-org/xplainfi}
  \item Report bugs at \url{https://github.com/mlr-org/xplainfi/issues}
}

}
\author{
\strong{Maintainer}: Lukas Burk \email{cran@lukasburk.de} (\href{https://orcid.org/0000-0001-7528-3795}{ORCID}) [copyright holder]

}
\keyword{internal}
