% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplainvi-package.R
\docType{package}
\name{xplainvi-package}
\alias{xplainvi}
\alias{xplainvi-package}
\title{xplainvi: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jemus42.github.io/xplainvi/}
  \item \url{https://github.com/jemus42/xplainvi}
  \item Report bugs at \url{https://github.com/jemus42/xplainvi/issues}
}

}
\author{
\strong{Maintainer}: Lukas Burk \email{github@quantenbrot.de} (\href{https://orcid.org/0000-0001-7528-3795}{ORCID})

}
\keyword{internal}
